requirements.txt
azure-cognitiveservices-speech==1.37.0
blinker==1.8.2
click==8.1.7
colorama==0.4.6
Flask==3.0.3
importlib_metadata==7.1.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
python-dotenv==1.0.1
Werkzeug==3.0.3
zipp==3.19.0
